findbugs BX_UNBOXING_IMMEDIATELY_REBOXED
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@22773 0c594bb0-8543-0410-a25e-d8ec440261db

+++ CefDataSource/src/org/autoplot/cefdatasource/CefMetadataModel.java
@@ -35,7 +35,7 @@ public class CefMetadataModel extends MetadataModel {
         if (o instanceof Float) {
             return ((Float) o).doubleValue();
         } else if (o instanceof Double) {
-            return ((Double) o).doubleValue();
+            return ((Double) o);
         } else if (o instanceof Short) {
             return ((Short) o).doubleValue();
         } else if (o instanceof String) {
